requirements.txt
requests>=2.28.0
pyinstaller>=5.0.0
pycryptodome>=3.18.0
ecdsa>=0.18.0
